Added two new environment variables used to configure the server and agent to allow more control over where pipeline directories are created.
The first environment variable, GO_PIPELINE_WORKING_DIR, is used by the
server to change the path where pipelines are create. If empty,
pipelines are created in the sandbox directory on the agent.

The second environment variable, GO_AGENT_SANDBOX_DIR, is used by the
agents to determine where the pipelines folder will be located.

Using these together you can set the GO_PIPELINE_WORKING_DIR to nothing
and then the GO_AGENT_SANDBOX_DIR to a scratch disk such that you
dedicate a drive for your go work.