Ignoring JS/CSS compression the right way.
The previous commit to ignore compression was alright. Except for the fact that merging was not considering a use-case where a library author might not end his/her JS file in a ';' and assumes that it will be loaded stand-alone. When we try to merge it into a single file, it throws up because of a malformed JS file. Fixed it by writing a ';' after the content is written if compression is switched off.