#000 - Bring back Go's override of Rails' url_for.
This override is used to replace the URL with the site URL and secure site
URL whenever they are defined. These tests, when ported over to Rails 4
failed. This is due to the fact that the root route in the old Rails was
set to {:controller => 'java', :action => 'null'}. So, all these tests,
when they used url_for, they got '/' as the path of the URL. In the new
Rails, it is 'welcome#index'. So, the routes specified in test_controller.rb
take over.