#000 - Tests were not cleaning up after themselves.
As a part of this test, the USE_COMPRESSED_JAVASCRIPT flag was wrongly set
to true, causing other tests to wrongly use compressed JS and succeed. At
least for Rails 2 codebase, this was wrong and the tests passed just because
of the order in which they ran. This commit makes sure that the test cleans
up after itself.