#620 - Fix resources save of 'RunOnAllAgents' job.
Since a resource config from the actual config object was being reused across multiple job
instances, only a single copy of the resource was being saved into the DB. Since the saving
was moved from iBatis to Hibernate, the save semantics changed. Unlike iBatis, Hibernate
realizes it is the same resource being shared across instances, and saves only one copy.

The fix is to send a different copy of the Resource object to Hibernate, for every save, so
that the save semantics mirror iBatis'.