#749 - Use debug logs for full exceptions.
Looked at hierarchy of calls, and changed error level logs to show only a message, but debug level logs shows the full
stack trace. Made this change in places which are on timers (somewhere up the stack), or are dependent on the number of
pipelines/jobs/properties in the config.