Make Keystore password configurable by ENV
Before creating the AppServer instance, it checks for the environment variable KEYSTORE_PASSWORD. If there is a password found, it uses that, if not (backwards compatible) it uses the hardcoded value stored 'password'. This should allow folks to change the password on the keystore for security sake where needed, just add the environment variable to /etc/default/go-server (or the Windows equivalent) - while maintaining compatibility with those who are fine with the historic hardcoded value.