Add a new TFS SDK 12.0.2.
There is a SystemProperty `tfs.sdk.old` that can be turned on
to fallback to v10.1.0 of the TFS SDK in case of issues.

The old tfs-impl module has been renamed to tfs-impl-10

There is a new copy of tfs-impl-12 that is identical to
tfs-impl-10, with a few changes to addresss differences
in TFS SDK versions.

The `agent.jar` contains 2 separate jars containing
`tfs-impl-10` and `tfs-impl-12` for the two implementations.
The `TfsSDKCommandBuilder` uses the `tfs.sdk.old` to determine
which of the two jars to load up for the TFS implementation.