Allow agent to switch between TFS 10.1.0 and 12.0.2 -
This is one elaborate haxx -

The server has a system property which when enabled, will
fallback to using TFS version 10.1.0.

Configuring agents to use a specific TFS version 
using system properties will be a huge hassle on 
large environments.

Therefore, on the server side, go will serialise 
information about the TFS version that the agent 
should use, the `TfsMaterial` implementation will
then setup the same system property as the server
so that the `TfsSDKCommandBuilder` will initialise
and load the correct TFS implementation version
(along with the TFS version)