Created a meaningful error message and error handling when files are inaccessible for an MD5 has computation.
Changes file access right checking to the create function rather than having it in the construct.

Also added absolute path folder into the error message should it occur.

Added test to test if the folder is displayed in the error message.

Removed unused imports.

Fixed fault in the logic when choosing which directory to display. Was the other way around. Fixed test as well.