#1133 GoConfigFileHelper should use config for edit, not current config
I think this change is not breaking anything, it looks like
GoConfigFileHelper is just a little out of date, being only a test
helper.
GoConfigFileHelper called dao only to get a config instance which it
then modified and serialized. Normally it does not happen like that,
there is a separate instance for edits.
I had to do this so that xml writer does not attempt to serialize a
merged config - which should never happen.