Remove confusion when creating `ArtifactPlan` instances.
Meat of the changes in `ArtifactPlan` and `TestArtifactPlan`. The rest
of it is supplimental changes to adapt to the change in their interfaces.

ArtifactPlan has a configurable `artifactType` instance. It was therefore
possible to create an instance of `ArtifactPlan` with `artifactType` set to
`ArtifactType.unit` - this was not only confusing, but incorrect since the
behavior of both artifact types is different.

This PR changes that by ensuring that the `artifactType` instance cannot
be set by clients that create an instance of `ArtifactPlan`.