Define X-UA-COMPATIBLE meta tag for IE intranet go sites
When the GoCD server is hosted internally, IE treats it as an intranet site by
default. Being an intranet site, it is automatically rendered in IE7
compatibility mode, which causes a few problems:

* pipelines dashboard doesn't line up as in Chrome and Firefox
* console log in job details page does not show line breaks, making it very
  difficult to read

By providing the X-UA-COMPATIBLE meta tag, IE shows the site in standards mode
and everything looks very similar to Chrome and Firefox.