#1762, #1779 - Bring back part of bug for backward compatibility
Earlier a user who was not in the 'operate' part of 'authorization' for a pipeline group was
allowed to be a stage approver. Changes made for #1762 (in #1779) removed that ability. But,
it is needed for backward compatibility for now. Otherwise, users' configs will be marked
invalid, without any change and without any notification to them. Will cause problems during
an upgrade.