#1133 fix invalid cached merged config in failed merged case
We want to ensure that UI changes are always present in current, valid,
cached config. There is always some valid base config. When partials are
involved there is least "last known valid merged configuration". So
regardless of the fact if latest main config and partials are mergable,
we can always fallback to using old partials with new modified main
configuration.