BuildCause in not kept in memory as a part of workAssignments, moreover it is not serialized over to the agents
BuildCause can be a huge object when there are too many modifications involved or even when a small number of modifications have a huge number of modifiedFiles in them. We seem to be loading up this object in memory for each job. When a given pipeline instance has a huge number of modifications/modifiedFiles, and the scheduled stage has too many jobs, this causes an out of memory of the agent. As the size of these modifiedFiles increase, it also tends to bring down the server. We do not need modifiedFiles to be passed along to the agents. We certainly don't need all of buildcause on the agent or in server memory as part of work-assignments, so removing that off. - #2145