Preventing duplicate modifications from getting inserted to database
When the scm history is re-written or a commit is merged without rebase, the scm-pollers do return duplicate revisions. These revisions are blindly inserted to db as new modification, even though these revisions were already part of db. This causes pipelines to trigger with the revisions that they had already triggered with and a whole lot of other confusion. This commit causes the duplicates to be ignored while inserting the modifications to db. If all revisions in a given MDU cycle for a material happen to be duplicate, then an exception is thrown to indicate the same - This would normally happen when the poller incorrectly keeps returning the same set of modifications all the time.

Adding a db constraint for materialid + revision makes sense, but it would require data correction which could get very complicated.
This prevents a bug in any given material-poller from screwing up the db with numerous duplicate modifications, and thereafter causing issues for everything else that relies on modifications data from db, ie. scheduling etc