Adding validation of config repo partials as a part of all config save flows
As of this commit, new Entity config APIs will validate against the valid-partials alone.
Throwing away PipelineConfigurationCache as the based on the partials configured, and changes done to it the cache will have to keep changing. Instead of caching this info, the required info is computed by PipelineConfigSaveValidationContext itself which would always have the right version of the cruiseConfig