Add a new Microsoft Team Explorer Everywhere (TEE) SDK 2015 v14.0.3
** THIS COMMIT BREAKS FUNCTIONALITY **

(From https://www.microsoft.com/en-us/download/details.aspx?id=47727)

There is a SystemProperty `tfs.sdk.old` that can be turned on
to fallback to v10.1.0 of the TFS SDK in case of issues.

The old tfs-impl module has been renamed to tfs-impl-10

There is a new copy of tfs-impl-14 that is identical to
tfs-impl-10, with a few changes to addresss differences
in TFS SDK versions.

The `agent.jar` no longer packages the tfs-impl jars. The sdk (in a
later commit) will be downloaded by the agent in a manner similar to
other jars (`agent.jar`, `agent-launcher.jar`) The server may choose to
send any any of the tfs-impl jars based on a system property.