Run integration tests using a separate task
... by moving them into a separate gradle `SourceSet`

Integration tests, as they are currently setup, run with all the unit tests.
Because the integration tests do not cleanup state at the end of each
run, they require the JVM to be forked on every run. This causes tests to
run really slow.

This PR sets up two different source sets, and 2 separate tasks to run
tests -

* `./gradlew server:test` - to run unit tests
* `./gradlew server:integrationTest` - to run integration tests
* `./gradlew server:check` - run all tests