Pass resolver which specifies the key to be cached for the given agent
Not passing key to the memoize method will take object reference as key by default.
Whereas on each poll we create a new agent object, and the memoize treats the newly created agent object as a new agent and will cache it seperately.
This was causing memoize to cache all the objects leading to memory overflow.