Added migration to handle removal of whitespaces from encrypted strings in config.
As a part of 17.2, we moved to Base64 implementation in Java 8 for encoding secrets from the one in commons-codec. The one in commons-codec would ignore whitespace characters, this allowed encrypted values to have newline, carriage-returns and spaces in an encrypted-value. Base64 impl in java8 is strict about this check and hence Go server fails to startup after an upgrade from 17.1 to 17.2 if the config had encrypted values with whitespace characters. This commit handles migration of such configs to remove the whitespaces.