Don't try to handle network disconnect for console log websocket connections
Sometimes a clean close(1000) initiated by the server is mischaracterized as an abnormal close(1006), at least
in Chrome, which causes the browser to continually reconnect and fetch the data over again. This seems to be a
race condition as a server initiated close(1000) is properly detected if there is a pause (e.g. from a debug
breakpoint). As we can't reliably detect a clean close(1000), don't try to handle a real disconnect (which
sends a 1006 close code anyway). For now, we will only handle 4004, which is our defined code for the console
log file not existing yet.