Add ability for LogOutputTransformer to defer lifecycle callbacks as well
Deferring lifecycle callbacks ensures in-order execution of deferred operations
when calling dequeue(). This is done by generifying transform() to invoke(), which
either executes or defers a function call with args; transform() becomes a mere
special case/syntactic sugar of invoke().

This should incidentally fix https://github.com/gocd/gocd/issues/3367 because the
detection for collapse/expand section happens in a lifecycle hook; when processing
is deferred because the console tab is not active, the detection mechanism will
now run upon dequeue().

[gocd_sf_team/#43] Console log expand/collapse all link does not always show