Always show collapse/expand all control
The detection logic is really only there to accommodate old logs that didn't have annotated
lines. The detection on a whole is probably not all that necessary, so opt for a simpler
implementation. I think it will be obvious to users that the link does nothing on older logs,
as those lack the collapse caret widget in the left gutter.

Note that this does not nullify the need for the prior commit (defer lifecycle callbacks); that
change is the proper way to handle lifecycle callbacks anyway.

[gocd_sf_team/#43] Console log expand/collapse all link does not always show