Fix a bug where single-line sections could potentially cause the preceding section to adopt its classes and error state
Discovered this edge case while implementing annotations for BuildSession. The order in which the job status lin was
outputted was slightly different than the old implementation, revealing this logic error. This is caused by the fact
that detectStatus() was invoked prior to the section boundary detection logic.

As detectStatus() is responsible for marking sections as errored, as well as assigning the right CSS classes to the
(current) section, if a job status line was preceded by a section that does not have an explicit boundary annotation
(e.g. prep or artifact publish sections), the call to detectStatus() will incorrectly mark the precedin section as
errored as well as assign it the CSS class to style a job status line.

This fix adds a check to see if the current line is an explicit boundary so that single-line sections, such as job
status, will be created, referenced by the current section pointer so that detectStatus() will decorate the intended
section.

That's a long commit message for such a tiny commit.