ConsoleLogEndpoint sends keepalive frames
@ketan provided an example of when this would be useful:

A user sets up a reverse proxy in front of Go, with a relatively short proxy read/send timeout. When tailing
the logs of a build, if a long-running task does not generate output within the read timeout, the proxy will
close the connection, and no further console data will reach the user.

To fix this, we send periodic keepalive frames during idle periods:

  - keepalive threshold is configurable via teh `go.websocket.keepalive.interval` property; defaults to 5000 ms
  - implemented as a PING frame, and only sent when no frames have been transmitted for a duration greater than or equal to the threshold