ConsoleLogEndpoint waits until the log file exists
Well, for a reasonable time anyway. Actually, it just retries 20 times until the file
exusts or until the connection is prematurely closed (browser refresh, user navigates away,
or network drops). If it doesn't exist after the last retry, ConsoleLogEndpoint closes the
connection and expects the browser to handle reconnection.

This situation can happen when a user is tailing the log and the job is still waiting to be
assigned to an agent. Solution is not ideal, but greatly reduces the frequency at which
connections are retried.

I chose 20 retries arbitrarily with a 500 ms interval to approximate 10 sec in most cases.
This will vary of course, especially with network filesystems, but it should be fine most
of the time.