Implement timeout detection in the browser
`SocketHealthService` on the server sends a ping message every 10s. The
JS in the browser keeps a tab on when it received the first ping event
to determine if the connection is still alive, and retry accordingly.

If the first websocket connection could not be opened, it is assumed
that a proxy in between is preventing websocket connections and we
fallback to the old AJAX polling mechanism.

Handle websocket connection close due to purging of console.log files