Simplify how agent.jar output is copied to file.
The current implementation does the equivalent of:

`java -jar agent.jar > go-agent-stdout.log 2>go-agent-stderr.log`

It does this by messing around with log4j NDC and filters which are
configured programatically, and any new appenders initialized later by
changing log4j.properties at runtime are not configured.

The new implementation, simply gets around this problem by directly
writing to a `RollingFileAppender` without going through the rest of the
logging framework.

Logs are optionally written to a `ConsoleAppender` if a system property
is set.