Silence some exceptions on server startup
(Addresses parts of #3754)

On startup the server is sending the plugin settings config and view
messages to all plugins to build plugin info objects. This call blows up
for atleast the task and authorization plugins because these calls are
not supported. Not making these calls will silence some of the
exceptions on startup.