Improve checksums for agent jar(s) download
Before this commit the (md5) checksums for various jars are were encoded
in base64 and hex encoding:

    curl -s https://example.com/go/admin/latest-agent.status -I | grep MD5
    Agent-Content-MD5: xcOIJGRG6DVEndmh+ngL0w==
    Agent-Launcher-Content-MD5: YZZ/8vOLt9tme93YVwtoIg==
    Agent-Plugins-Content-MD5: 0fc9e52a2619dedfd7973959b8c92063
    TFS-SDK-Content-MD5: FDZDpIK84xCKx0EO+HZu9A==

This caused problems when checksums were verified because the agent was
always verifying passwords using base64, and the plugins zip was encoded
in hex format, causing the agent to re-download the plugins.zip every
time it was started up.

Additionally there was another bug with how the `agent-plugins.zip`
checksum was computed. The checksum was computed was basically the
equivalent of:
`cat [list of plugin.jar] | md5sum`.

Whereas it should have been:

`zip [list of plugin.jar] | md5sum`