Fixed a race condition between rescheduleHungJobs and rescheduleAbandonedBuildIfNecessary (#3891)
Both these threads invoke rescheduleJob. The threads coud have run the queries and gotten the same jobid from the corresponding queries, one of them acquires a lock first and takes care of rescheduling the job. The other thread would wait for its turn, and then enter the sync block and try to do the same causing a rescheduled job to be rescheduled twice. While this can happen for legacy agents too, its seen more often when websocket is turned on