bringing back set_error_flash as templates_controller still uses it. This was missed while migrating to rails4 and caused the destroy action to be broken in certain cases
Currently a javascript check prevents a user from clicking the delete button if the template is associated with a pipeline. However, if user1 lands on the template page before its associated to a pipleine and clicks on the delete button, lands on the confirmation popup. meanwhile user2 associates a pipeline with this template. Then if user clicks on the proceed button, the application throws up a 500 page.