Fix issue with whitespace (#3966)
The current implementation of
`AgentProcessParentImpl#agentInvocationCommand` will look at
`AGENT_STARTUP_ARGS` env and split those by a whitespace which causes
an incorrect command line to be assembled.

We now set the log dir in `agent.sh` only in cases where the agent is
invoked via init script. When executed via `./agent.sh`, the default log
dir is `$PWD/logs` and need not be passed in.