Introducing token based approach for agent registration
* added new endpoint to get token from server in exchange of agent uuid.
The token for a given agent is issued for unknown agents only - already
registered or pending agents cannot request for a token.

* agent requests for token before registration if required. This token
is stored on agent's disk and reused for registration calls. Storing to
disk is required to handle agent-process restarts during registration
and in cases where the agent's certs are deleted and a fresh
registration call is made using the same token.