Separating out JVM arg used for session cookie max-age from session max-inactive interval (#4099)
Now that both of these are configurable via the same jvm arg (ie. go.server.session.timeout.seconds), an enduser could potentially set cookie's max-age to -1 to make the cookie expire as soon as the browser was closed - which is good. But this leads to (mostly unintentionally) setting the session's max inactive interval to -1 which causes a session to never expire. So, even if upon a cookie expiry a new session gets created for the user, the old session remains on the server adding a memory overhead.

Added a separate JVM arg to configure session cookie's  max age using go.sessioncookie.maxage.seconds.  This defaults to 2 weeks, could be set to -1 to expire the cookie on browser close. Should not be set to 0 else the cookie will expire immediately.

No changes done to go.server.session.timeout.seconds. Users must never set this to -1, unless ofcourse they know the repercussions