Delete entities associated with a job
When a job completes, a lot of entities are left behind in the DB. These
persisted entities are copies of the config at the time the job was
scheduled, and is not used once the job completes.

For jobs that are already completed (as per DB), and for those jobs that
will complete from this point onwards:

* delete environment variables that are not trigger-with-options
* delete all artifact plans that are of type `file`. Artifact plans of
  type `unit` are retained for shine.
* delete all artifact properties generators
* delete all (agent) resources associated with a job
* delete all elastic agent metatada

Since some of the data in the DB will eventually go missing, the job
feed api will break. We will no longer render the environment variables,
resources and build artifacts in the job xml api.