upgrading objenesis
Noticed about 150 threads blocked at synchronized method ObjenesisBase.getInstantiatorOf in objenesis v1.2 for one of the performance runs. This issue was resolved in v2.0 (method is not synchronized anymore), but upgrading to latest available (v2.6) as of now. The least we need is 2.0, so if we notice an issue with 2.6, we could try downgrading to 2.0