New dashboard did not list pipeline instances with failing stages
A stage is marked as failing when all the jobs were building/scheduled and one of those jobs fails.
The query for the older dashboard doesnot load up such stages too, however, it does show up pipeline instances with failing stages if the failure happened after the server was brought up (some cache data magic). If the server was restarted with db being in such a stage, even the old dashboard did not list such instances.
This is an issue since a running stage prevents running of the same stage in a different pipeline instance. Without the pipeline instance being visible, its difficult to spot these kind of situations. Even if this behavior changes in future, it is important to see all running instances on the dashboard.