Got rid of PipelineSqlMapDao.findPipelineByCounterOrLabel
The lookup was buggy since it assumed a numeric input was always a counter. By default label=counter, so if parent pipeline left labelTemplate to be default, but child pipeline set it to be revision of parent pipeline, then this method would return incorrect results. Also, there is not requirement of looking up a pipeline by its label.