Making pipeline-name in PipelineLabelCounter case-insensitive
PipelineName in PipelineLabelCounter table was case-sensitive which causes the pipeline-counter to reset if the pipeline is renamed with just a different case.
This commit adds a PipelineLabelCorrector which deletes the duplicate entries from PipelineLabelCount table. The entry which matches the case of the pipeline name in config is left behind.
It introduces another column caseInsensitivePipelineName which is a case-insensitive column. All queries have been updated to make use of this column now
This fixes a few issues:
* Pipeline counter doesn't get reset upon change of case of pipeline-name
* Renaming(changing case) of a building pipeline used to show up as never-run on new-dashboard. It shows the existing state now.
* #1471, #3055 - These wouldn't happen again. Although the PR doesn't do anything to fix the data for existing pipeline runs. The issue wouldn't happen for future runs
* Fixed bug with pause/unpause API wherein it would incorrectly report that the pipeline has been paused but wouldn't really pause it if this was invoked with a different case

There might be other issues in the application  with case-insensitivity of pipeline,stage and job names. This commit doesn't claim to fix all of those