Fix broken analytics drilldown mechanism by namespacing models and managing them in one place
Context: (i.e. why such a big change to fix this?)

`PluginEndpointRequestHandler.defineFetchAnalyticsHandler(models)` was being called twice in a page; while
this is not bad in and of itself, it was called with a different `models` reference, which ended up clobbering
the model cache, thereby causing model lookups to fail when serving `fetch-analytics`.

Sure, I could have just removed the second invocation, and called it a day. However, I felt this was a deeper
problem with the design.

The `defineFetchAnalytics()` method was called within the context of a widget/component that dealt with rendering
charts for the analytics dashboard. Naturally, one would want to ensure all dependent mechanisms were set up, so
it's understandable that one would think to bootstrap all of this within that widget. It's not obvious that this
would cause failures everywhere else.

My point is, the design of how these components talk to each other, and where responsibilities are split lended
itself to breaking the app, and we should fix the design so that things like this don't happen so easily.

Now that that's out of the way, here are the changes:

1. Introduce `AnalyticsInteractionManager`

  - Responsible for:
    - Setting up listeners: `AnalyticsInteractionManager.ensure()`
    - Managing models within a page.

  - Model management is hidden and centralized so there is no way to accidentally overwrite
    the cache. However, it is also namespaceable so that widgets only need concern themselves
    with the models they manage. A good example is the analytics dashboard; global charts don't
    concern themselves with pipeline charts, and vice-versa.

    > `AnalyticsInteractionManager.ns("SomeName")` yields a `Namespace`
    > object.

  - Model/iframe `uid`s are now standardized, and the `Namespace` objects will generate them
    for you.

      - `uid`s are predictable and deterministic (calculable from the given parameters). In
        fact, they can be decoded back into the source information, which makes it convenient
        to build URLs and do other tasks that are specific to a plugin.
      - `Namespace` abstract model construction via the `modelFor()` method. It returns a
        preconfigured model for a particular iframe.

2. Split out `PipelineDropdown` component from `PipelineMetrics`

3. `Frame` no longer takes a callback during construction. `Frame.load(before, after)` now takes
   hooks to fire before and after data load, which is convenient for managing things like spinners
   and calls to `m.redraw()`.