Optimization of `GoDashboardCurrentStateLoader`
Do not perform the DB query to load pipelines if list of pipelines has
not changed.

A further optimization would be to load from the DB, only the pipelines
that have been newly added. Pipelines that have been removed do not
require a DB call, and we can simply purge those from the
`historyForDashboard` instance variable.