#4874 - Revert d416b47ce18bb15fc54bebcf7fe8ae61c0de027e
Brings back "Eager load a quick request handler" functionality.

The Jetty Server is started empty, with no handlers and that's quick. It takes ports 8153 and 8154 and starts returning
404 for everything. Immediately, the welcome file handler is started. It starts handling / (redirecting to /go/home)
which won't work, because the Spring web app (and Rails) haven't started yet. Once they're up, the redirection works.

Till that time, we can show a message or something more clever, while waiting for the rest of the app to come up.