Fixed issue of duplicate html elements showing up on new dashboard - #4775
Foundation components were being used for buttonWithTooltip and linkWithTooltip. While these worked in most cases, when such elements are present of pages which get refreshed by some ajax call (as was the case on dashboard), the destruction of these foundation elements (if the element was to be removed from the UI) wasn't clean and lead to issues. The first call to component.onremove would succeed but mithil kept a reference of the old/deleted element and called onremove on it all over again leading to onremove call failing. Since oncreate was being called before onremove, the element wasn't really deleted from the UI. Using a custom html component instead to solve this.