Continue oauth flow when security is disabled and oauth credentials provided in request.
We allowed oauth request to continue when security is disabled and oauth
credentials are provided in the request.

When security is disabled
[this](https://github.com/gocd/gocd/blob/1013a67ccf0cda9c87e6361286708b584be42338/server/src/main/java/com/thoughtworks/go/server/newsecurity/filters/OauthAuthenticationFilter.java#L105)
will block user if oauth credentials are present in the request.

BC addons uses oauth flow for to sync dataa from primary server and
above change will affect the BC addon workflow.

- When security is disabled BC aquires a Oauth token from primary server
and uses it to make sync requests hence primary reject the request as
oauth token present in the request.

Primary server should note generate oauth token when security is
disabled because oauth token is for a user and without security there
will not be any user. I am leaning towards not allowing BC to work when
security is disabled.