Data sharing and update check notifications
Create a separate UI component for showing up notifications
This is now used for showing the notification for data-sharing and also for displaying the notification for newer releases.
The notifications are stored in the browser local-storage
Users can mark the notification as read and wouldn't see them again on that browser.