Migrate PipelineSelections to a JSON format
  - Add a schema version column and a CLOB column to hold a JSON doc to describe the filters
  - Add data migration (and mini framework) to transduce current CSV format to JSON

The data migrations run as a task after dbdeploy runs its migrations. Hopefully, this will be only a
temporary workaround until we move away from dbdeploy (which does not allow Java migrations). This is
the reason we need a version column in the pipelineselections table; it allows us to determine whether
or not a set of migrations need to be run or not against data that has a flexible structure, such as
JSON.

Alternatively, we could have introduced a new table to track which data migrations have already run.
This is probably more reliable and efficient in the long run; on the downside, it introduces yet
another schema version table and increased complexity. I'm also being optimistic that we will eventually
move away from dbdeploy, so I'd hate to introduce a second table just for data migrations.

The data migration will essentially convert the pipeline selections representation into a JSON document
with this structure:

```
{
  filters: [
    { name: "first filter", type: "whitelist", pipelines: [ "p1", "p2", ..., "pN" ] },
    { name: "second filter", type: "blacklist", pipelines: [ "p4", "p5" ] },
    ...
  ]
}
```

The implications here are that we can support multiple pipeline filters per user as a single row in the
table and that this will allow us to consolidate and shift more logic to the frontend. The backend can
be written in a way such that its only concerns are serialization, deserialization, and persistence.