Revert "Add the main jar to the classloader"
This reverts commit 1fc6911b4769af82ae81e27a3e74ab0dff976d66.
Many of the threads were blocked waiting to acquire a lock after entering the catch-block. The catch block did a  which internally has a synchronized block causing this. Given main.jar was being looked up for all dependency lookup, it kept going to the catch block while looking up for classes from other dependencies. Changing the order in which urls are specified for UrlClassLoader and pushing main.jar as the last item still doesn't fix things. Hence reverting the commit.