#5020 - Fix request logging.
Jetty server level handler is now a HandlerCollection which wraps the ContextHandlerCollection used by
the DeploymentManager to handle the loading indicator. Since the RequestLogHandler added by jetty.xml
will now be at the Jetty server level, both the ContextHandlerCollection and RequestLogHandler will be
called (one after the other). So, request logging works.

If the ContextHandlerCollection was not added into the HandlerCollection, the RequestLogHandler would
be added into the ContextHandlerCollection. Then, the RequestLogHandler wouldn't be called, since the
ContextHandlerCollection is special and calls only the handlers whose context path matches the request.