Honor expanded/collapsed group states in personalization modal
We need to keep a static model of the complete list of groups and pipelines and their various states.
`PipelineListVM#displayedList()` should return a subset of this list upon searching, still holding
references to the original streams within the static model.

Additionally, track 2 distinct states for expand/collapse per pipeline group: a state during search and
a state when there is no search term. This allows us to restore the expanded/collapsed states of each
group after the user clears the search term.

The implied behavior is that during a search, all matches are expanded by default, but the user can
freely customize the expand/collapse states of the search results  without affecting the original
group states prior to the search.