Perform backup via a cron defined in config xml
This commit just covers the XML config along with a timer service. APIs
will be in a later commit.

Introduce a `<backup />` tag under `<server />` that will allow users to
configure the timer, a post-backup hook script, and flags to indicate if
an email should be sent on backup success and failure.

The script will be invoked regardless of whether the backup succeeds or
fails, and will receive the backup details, who triggered it, and if the
backup was successful. This should allow sufficient flexibility for the
script to make decisions as appropriate.

A server health message will be populated on backup failures (or
post-backup script failures). This message will not be persisted across
restarts, which is the current behavior with every server health
message.