Change `GoConfigRevision` to be backed by `byte[]`
`GoConfigRevision` represents a commit in the config repository along
with the contents of the config XML at that revision.

The implementation of `GoConfigRevision` suffers from a performance
issue in that the content of the config XML is currently converted from
a `byte[]` to `String`. This can take a few seconds for larger
configurations. An instance of `GoConfigRevision` is created on every
admin page (for the md5 check), and this conversion is unnecessary when
the content is not really needed. Ideally the content should not be
present in the `GoConfigRevision` and should be loaded only when the
content needs to be rendered. This is probably an optimiation for later.