Making a bunch of APIs to support pipeline-counter instead of both counter and label (#5256)
* Avoiding db query to check for pipeline and job when artifact is being uploaded by an agent

* Updated artifact APIs to accept only pipeline counters (not labels)

None of the client (javascripts, go-agent, APIs) pass along pipelineLabel to the artifacts API, but it still for some reason handled it adding complications and unnecessary db queries to get the label

* Stages rerun, properties and job details page do not accept label now. All clients were passing along pipeline counter anyhow

* fixing tests

* Fixed bug in properties controller to handle 'latest' for pipeline and stage counters

* Updated stages rerun code to make use of pipeline lookup by pipeline counter(instead of label).

* Got rid of all usages of PipelineService.findPipelineByCounterOrLabel

* Stage locator never has pipeline label, hence StageIdentifier need not handle it

This unnecessarily compiled a regex on every call and always resulted in no-match

* Got rid of StageService.findStageIdByLocator - couldn't find any usages in production code

* Got rid of PipelineSqlMapDao.findPipelineByCounterOrLabel

The lookup was buggy since it assumed a numeric input was always a counter. By default label=counter, so if parent pipeline left labelTemplate to be default, but child pipeline set it to be revision of parent pipeline, then this method would return incorrect results. Also, there is not requirement of looking up a pipeline by its label.

* getting rid of the test to handle 'latest' in Job details page url since we do not use that from anywhere in the application. Moreover, it did not fully function

* Removing some unused code and related test

* Move duplicate function into PipelineService

* Fix tests and use a better name for the method resolving pipelineCounter

* Fix SQL syntax errors