Basic cut of config repo SPA
- lists all config repos
- allows opening modal to edit

Added a new `ApiRequestBuilder` to replace AjaxHelper and friends, it
simply wraps `m.request` and returns the `XMLHttpRequest` object in the
promise. The deserialization and everything is left to the caller. See
`ConfigReposCRUD` for example.

For GET calls that need to capture etag+response object and PUT calls
that need to submit these, introduced a `HttpResponseWithEtag` object to
encapsulate these.

TBD:

- actually saving after editing
- error handing
- creating config repo
- deleting config repo