Ensure the PipelineConfig hierarchy computes nice `hashCode()`
Made sure to include any config subtab entries in the `hashCode()` computation
so we can rely on these for uniqueness.

This will help us compute an etag for pipeline export, which will depend on
pipeline config, plugin id, and plugin version (the latter 2 will be really
a hashCode() of GoPluginDescriptor).