Add GoPartialConfig.merge() as a convenience to merge a PartialConfig into a CruiseConfig
To achieve this, we also extracted the inner instance class GoPartialUpdateCommand into a top-level
class, PartialConfigUpdateCommand. This required abstracting some behavior behind a new interface,
PartialConfigResolver and the implementation was moved to CachedGoPartials as it seemed more appropriate
given the fact that all of the logic only concerned the partial caches.