Add `GoConfigService.validateCruiseConfig()` as a convenience method to perform validations
This required some significant refactoring to GoConfigInvalidException and GoConfigInvalidMergeException
so that we can access the errors as a list from the exceptions directly. This will help build the preflight
result that will be returned by the API, as it is generally more useful to retain a list of errors rather
than a concatenated error message. The `summary` string was also dropped from the exception constructors
as the value wasn't being used(!) in code anyway.

MagicalGoConfigXmlLoader now exposes `validateCruiseConfig()`, which is delegated to by GoConfigService.