Remove dependence on some rails gems
`rails` was an "empty" gem that merely dependended on a set of gems to
offer a bunch of functionality for webapps. Some of this was unsed by
gocd, and was continued to be packaged along with the installer, along
with their transitive dependencies.

This PR gets rid of `rails`, specifically to remove the following unused
gems:

- actioncable
- actionmailer
- activejobs
- activerecord
- activestorage