Handle plugin uninstall gracefully.
- Should not uninstall the plugin if it is already uninstalled.

On installation of a plugin, GoCD server will load it and cache the file system information.
When a plugin is identified as invalid, server will stop and uninstalls it from OSGI but file
system cache is still there(this is needed to detect whether plugin is added or updated). So
on update of the same plugin jar system tries to unload it again and throws an error as it was
previouly uninstalled. This check will see if plugin bundle in uninstalled or not and based on
that handles further operations.