[#5614] If plugin is incompatible with current gocd version, added error message stating the same. (#5801)
* [#5614] If plugin is incompatible with current gocd version, added error message stating the same.

* Validate plugin extensions versions with GoCD server supported extension versions.

- Plugin will be marked as invalid if any one of the extension version
fails to match the GoCD supported extension version.

* Updated test to use assertj matchers

* Mark a plugin as invalid if the plugin is implmentrd using unknown
extension(s)

* Adding ExtensionsRegistry bean declaration in agent's applicationContext.xml

* Handle plugin uninstall gracefully.

- Should not uninstall the plugin if it is already uninstalled.

On installation of a plugin, GoCD server will load it and cache the file system information.
When a plugin is identified as invalid, server will stop and uninstalls it from OSGI but file
system cache is still there(this is needed to detect whether plugin is added or updated). So
on update of the same plugin jar system tries to unload it again and throws an error as it was
previouly uninstalled. This check will see if plugin bundle in uninstalled or not and based on
that handles further operations.