Add a login page/logout page SPA
The new page can be toggled using a feature toggle. The page contains a
`data-meta` attribute to indicate if password/web plugins are installed,
and what web authentication plugins are installed to be able to
determine whether the login form should be shown and what the web auth
plugin images and links should be.